requirements.txt
python-telegram-bot==20.7
requests==2.31.0
matplotlib==3.8.2
numpy==1.26.3
